Awesome new feature: Google Docs Converter
So I was working with the Google Drive API, when I worked out it I
could get it to convert documents for me. So now we have 2 ways
to get the job done.

Both require setup, you either need to install unoconv on your system.
Or you have to create an OAuth Key in the Google Devlopers Console.

From my testing, there really isn't alot of difference between unoconv
and the Google converter. It wouldn't surprise me if they were using
something similar behind the scenes anyway.

However the Google Converter is obviously slower, having to upload and
download the document. But might be an option for shared / restrictive
environments.