Fixes the download and stream methods.
Before in the event that a simple basic conversion was performed and
then, streamed or downloaded to the browser. The generated PDF contained
a PHP exception. Because the "tempDocument" didn't actually exist.

We also changed the way we delete the temp documents.
This is done via a destructor now for the entire class.